Split time and distance measurers and use separate file for get_measurer
